# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Bitaqty Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/vendors/credits": {
            "get": {
                "tags": [
                    "VendorCredits"
                ],
                "summary": "Credits",
                "description": "Vendor Credits",
                "operationId": "VendorCredits",
                "parameters": [
                    {
                        "name": "filter[created_at]",
                        "in": "query",
                        "description": "Filter credits by created_at date range",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2018-01-01,2018-12-31"
                    },
                    {
                        "name": "filter[deposit_or_withdraw]",
                        "in": "query",
                        "description": "Filter credits by deposit_or_withdraw",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "filter[credit_type_id]",
                        "in": "query",
                        "description": "Filter credits by credit_type_id",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "filter[credit_status_id]",
                        "in": "query",
                        "description": "Filter credits by credit_status_id",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": ""
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "amount": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "notes": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "deposit_or_withdraw": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "credit_type": {
                                                        "type": "object",
                                                        "example": {
                                                            "id": "",
                                                            "name": ""
                                                        }
                                                    },
                                                    "supported_account": {
                                                        "type": "object",
                                                        "example": {
                                                            "id": "",
                                                            "name": "",
                                                            "image": ""
                                                        }
                                                    },
                                                    "credit_status": {
                                                        "type": "object",
                                                        "example": {
                                                            "id": "",
                                                            "name": ""
                                                        }
                                                    },
                                                    "from": {
                                                        "type": "object",
                                                        "example": {
                                                            "name": "",
                                                            "email": "",
                                                            "phone": "",
                                                            "image": ""
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/vendors/credits/request": {
            "post": {
                "tags": [
                    "VendorCredits"
                ],
                "summary": "Vendor Credits Request",
                "description": "Vendor Credits Request",
                "operationId": "VendorCreditsRequest",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image",
                                    "amount",
                                    "notes",
                                    "supported_account_id"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "file",
                                        "format": "file"
                                    },
                                    "amount": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "supported_account_id": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "errors": {
                                            "properties": {
                                                "dynamic-error-keys": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}